insights-1148, part 2 -move upload functions to new module so it can be called from future container module -run collection through multiple 'targets' specified by container module, default is original behavior
